Structure refactor, Token format change, Introduce Pipenv
I've done a refactor of the structure of the program, all the bot files
are now contained inside luhack_bot as a module.

I've also changed the token format, now the timestamp and user_id is
encoded inside the token & signed (I can remove this if it's not wanted)

All database interactions have been refactored to use query parameters,
instead of string formatting.

I've introduced pipenv to manage running the bot in a environment with
all it's dependencies ensured.